% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HPcluster.R
\name{HPcluster}
\alias{HPcluster}
\title{Hurdle Poisson clustering}
\usage{
HPcluster(data, Treatment, nK, method = c("EM", "SA"), cool = 0.9, nstart = 1)
}
\arguments{
\item{data}{Data matrix with dimension N*P}

\item{Treatment}{Vector of length P. Indicating replicates of different treatment groups. For example, \emph{Treatment} = c(1,1,2,2,3,3) indicates 3 treatment groups, each with 2 replicates.}

\item{nK}{Positive integer. Number of clusters.}

\item{method}{Method for the algorithm. Can choose from \emph{"EM"} as Expectation Maximization or \emph{"SA"} as Simulated Annealing.}

\item{cool}{Real number between (0, 1). Cooling rate for the \emph{"SA"} algorithm. Uses 0.9 by default.}

\item{nstart}{Positive integer. Number of starts for the entire algorithm. Note that as \emph{nstart} increases the computational time also grows linearly. Uses 1 by default.}
}
\value{
\describe{
\item{cluster}{Vector consists of integers from 1 to nK, indicating final clustering result. For evaluating the clustering result please check \link[aricode]{NMI} for \emph{normalized mutual information}.}
\item{prob}{N*nK matrix, the (i, j)th element representing the probability that observation i belongs to cluster j.}
\item{log_l}{Scaler, the hurdle Poisson log-likelihood of the final clustering result.}
}
}
\description{
This function gives the clustering result based on a hurdle Poisson model.
}
\examples{
HPcluster(sample_data, rep(c(1,2,3), 10), nK = 7, method = 'EM', cool = 0.9, nstart = 1)
}
