% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hybrid.R
\name{Hybrid}
\alias{Hybrid}
\title{Calculate optimal number of clusters.}
\usage{
Hybrid(data, Treatment)
}
\arguments{
\item{data}{data matrix}

\item{Treatment}{Vector of length p, indicating replicates of different treatment groups. For example, \emph{Treatment} = c(1,1,2,2,3,3) indicates 3 treatment groups, each with 2 replicates.}
}
\value{
A positive integer indicating the optimal number of clusters
}
\description{
This function calculates the optimal number of clusters for a given dataset.
}
\examples{
Hybrid(sample_data, rep(c(1,2,3), each = 10))

}
