% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\docType{data}
\name{sample_data}
\alias{sample_data}
\title{Sample of sparse microbiome count data}
\format{
The dataset contains 30 columns, indexed as A1 ~ A10, B1 ~ B10, C1 ~ C10 to represent 3 treatment groups.
}
\usage{
sample_data
}
\description{
A sample data frame with 300 features, 3 treatment groups with 10 replicates in each group.
}
\examples{
head("sample_data")
}
\keyword{datasets}
